skyobjet/ improve Root, LastRoot methods, improve documatation, fix holding
